+++ CHANGELOG.md
@@ -35,6 +35,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `snapshot.max-new-file-size` config option. It will print a warning and large
   files will be left untracked.
 
+* New command option `--config-file=PATH` to load additional configuration from
+  files.
+
 * Templates now support the `>=`, `>`, `<=`, and `<` relational operators for
   `Integer` types.